python2_requirements.txt
Django==1.11.13
django-celery==3.2.2
django-crispy-forms==1.7.2
celery==3.1.26.post2
kombu==3.0.37
amqp==1.4.9
ipython==5.6.0
ipython-genutils==0.2.0
Logbook==1.3.3
lxml==4.2.1
pathlib2==2.3.2
Pillow==5.1.0
psycopg2==2.7.4
python-dateutil==2.7.3
pyunpack==0.1.2
reportlab==3.4.0
suds==0.4
Unipath==1.1
python-memcached==1.59
pyinotify==0.9.6
django-registration-redux==2.5
#pycurl==7.43.0.1
patool==1.12
rarfile==3.0
qrcode==6.0
redis==3.0.1
#M2Crypto==0.33.0
openpyxl==2.6.2
#pykcs11==1.5.5
django-multiselectfield==0.1.8
architect==0.5.6
